Add ohlc_cache.go to create cache for opening and closing price data.
Discussion:

The free IEX site (iextrading) API has an endpoint for official opening prices, however after
investigation I was unable to get the endpoint to produce any data.

The secured, paid tier has a separate endpoint for each stock that produces this data point,
however it is obviously not secure to disseminate the API key to the client application.

The solution is to have the server call this endpoint for all stocks at specific times, cache
the data, then provide to the client through a separate endpoint that it manages.

The free tier api can provide the real time data.